LayoutTests/plugins/quicktime-plugin-replacement.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../resources/js-test-pre.js"></script>
        <script>
            var embed;
            var src = "resources/orange.mov";

            function logEvent(event)
            {
                debug("<i>EVENT(" + event.type + ")</i>");
            }
            
            function seeked()
            {
                debug("");
                shouldEvaluateTo("embed.GetTime()", embed.GetTimeScale() * 2);
                evalAndLog("embed.Play()");
            }

            function playing(event)
            {
                debug("");
                logEvent(event);
                shouldEvaluateTo("embed.GetRate()", 1);
                evalAndLog("embed.Stop()");
            }
            
            function paused(event)
            {
                debug("");
                logEvent(event)
                shouldEvaluateTo("embed.GetRate()", 0);

                debug("");
                evalAndLog("embed.SetURL('')");
            }
            
            function error(event)
            {
                logEvent(event)
                shouldBeEqualToString("embed.GetURL()", "");
                shouldBeEqualToString("embed.GetPluginStatus()", "Error");
                debug("<br>END TEST.");
                if (window.testRunner)
                    testRunner.notifyDone();
            }
            
            function test(event)
            {
                logEvent(event)

                if (window.internals) {
                    debug("<br>** Test to make sure &lt;video&gt; element is in the shadow DOM.");
                    evalAndLog("shadowElement = internals.shadowRoot(embed).firstChild");
                    shouldBeEqualToString("shadowElement.tagName", "VIDEO");
                    shouldBeEqualToString("internals.shadowPseudoId(shadowElement)", "-webkit-plugin-replacement");
                } else {
                    debug("");
                    testFailed("This test must be run in DRT.");
                    debug("");
                }

                debug("<br>** Test for the presence of overridden methods.");
                var methods = [ "SetURL", "GetURL", "Play", "Stop", "GetRate", "SetRate", "IsFullScreen", 
                                "ExitFullScreen", "GetPluginStatus", "GetTime", "SetTime", "SeekToDate", "GetDate", 
                                "GetDuration", "GetTimeScale", "GetMaxTimeLoaded", "GetMaxBytesLoaded", 
                                "GetMovieSize", "GetTimedMetadataUpdates", "GetAccessLog", "GetErrorLog"
                              ];
                for (i = 0; i < methods.length; ++i) {
                    var type = typeof embed[methods[i]];
                    if (type == "function")
                        testPassed("embed." + methods[i] + " is a function.");
                    else
                        testFailed("embed." + methods[i] + " should be a function but is \"" + type + "\".");
                }

                debug("<br>** Test overridden methods.");
                shouldBeTrue("embed.GetURL().indexOf('" + src + "') != -1");
                shouldEvaluateTo("embed.GetDuration()", 300000);
                shouldEvaluateTo("embed.GetMovieSize()", 24962);
                shouldEvaluateTo("embed.GetMaxTimeLoaded()", 300000);
                shouldEvaluateTo("embed.GetMaxBytesLoaded()", 24962);
                shouldEvaluateTo("embed.GetMovieSize()", 24962);
                shouldBeEqualToString("embed.GetPluginStatus()", "Playable");
                
                // "GetDate()" should return the current date, clear seconds and milliseconds 
                var now = new Date();
                shouldBeCloseTo("new Date(embed.GetDate()) - new Date()", 0, 100);

                evalAndLog("embed.SetTime(embed.GetTimeScale() * 2)");
                setTimeout(seeked, 500);
            }
            
            function setup()
            {
                if (window.testRunner) {
                    window.testRunner.waitUntilDone();
                    window.testRunner.dumpAsText()
                    window.internals.settings.setPluginReplacementEnabled(true);
                }
                div = document.getElementById("parent");
                div.innerHTML = '<embed src="' + src + '" width="640" height="480" postdomevents=true>';
                
                var events = [ "qt_begin", "qt_validated", "qt_enterfullscreen", "qt_exitfullscreen", 
                                "qt_loadedmetada", "qt_canplay", "qt_ended", "qt_timedmetadataupdated" 
                              ];
                embed = div.firstChild;
                for (i = 0; i < events.length; ++i)
                    embed.addEventListener(events[i], logEvent, true);
                embed.addEventListener("qt_canplaythrough", test, true);
                embed.addEventListener("qt_play", playing, true);
                embed.addEventListener("qt_pause", paused, true);
                embed.addEventListener("qt_error", error, true);
            }
        </script>
    </head>

    <body onload="setup()">
        <div id=parent></div>
        <p>
        <div id=console></div>
    </body>
</html>
